requirements.txt
Django==1.9.7
Pillow==3.2.0
PyJWT==1.4.0
django-cors-headers==1.1.0
django-sslserver==0.19
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
psycopg2==2.6.1
